requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
streamlit>=1.35.0
pandas>=2.1.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
langchain>=0.2.0
langchain-google-genai>=1.0.6
langchain-groq>=0.1.0
langchain-community>=0.2.0
google-generativeai>=0.8.0
plotly>=5.20.0
scikit-learn>=1.4.0
xgboost>=2.0.0
python-dotenv>=1.0.0
requests>=2.32.0
pydantic>=2.7.0
sqlglot>=20.0.0
asyncpg>=0.29.0
